Fix Memo ID Type

+++ src/app/src/main/java/jp/kaleidot725/emomemo/model/entity/Memo.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "memo")
 data class Memo(
-    @PrimaryKey(autoGenerate = true) val id: Long,
+    @PrimaryKey(autoGenerate = true) val id: Int,
     val tag: String,
     val title: String
 )
